fix some bugs, move zoomType and zoom_wit_mouse_wheel method in general.py
